relocate io.reader make test_io.test_base_writer more strict fix model.csv.GenericCsvModel     (this fix stop raising NotImplementedError) add metadata save func to io.writer refactaroring io.writer
