--- test/records/jera/test.csv.json
@@ -1,8 +1,0 @@
-{
-    "description": "forecast",
-    "database_id": "JERA Forecas",
-    "record_id": "test",
-    "type": "raw_data",
-    "path": "/opt/pydtk/test/records/jera/test.csv",
-    "content-type": "text/csv"
-}
