+++ test/test_io.py
@@ -21,7 +21,7 @@ def test_base_reader():
 
 def test_base_writer():
     """Run the base writer test."""
-    from pydtk.io import BaseFileWriter
+    from pydtk.io import BaseFileWriter, BaseFileReader
     import json
     import os
     import numpy as np
@@ -32,29 +32,34 @@ def test_base_writer():
         ["2020 / 11 / 03", "00:00", 2.7814903259, 4.0177507401, 4.795976162, 5.4033985138, 8.051410675],
         ["2020 / 11 / 03", "00:30", 2.6788742542, 3.86374259, 4.5651106834, 5.2544236183, 7.7826366424],
         ["2020 / 11 / 03", "01:00", 2.5762581825, 3.7097344398, 4.3342452049, 5.105448722799999, 7.5138626099],
-        ["2020 / 11 / 03", "01:30", 2.4808146954, 3.7635395527, 4.2475838661, 4.9490222930000005, 7.6304917336],
-        ["2020 / 11 / 03", "02:00", 2.3853712082, 3.8173446655, 4.1609225273, 4.7925958633, 7.7471208572],
-        ["2020 / 11 / 03", "02:30", 2.4931795597, 3.8574020863, 4.1738054753, 4.8718442916, 7.7336649895],
-        ["2020 / 11 / 03", "03:00", 2.6009879112, 3.897459507, 4.1866884232, 4.95109272, 7.7202091217],
-        ["2020 / 11 / 03", "03:30", 2.6634838581, 3.9712848664, 4.2355024815, 4.903665781, 7.9435429573],
-        ["2020 / 11 / 03", "04:00", 2.725979805, 4.0451102257, 4.2843165398, 4.856238842, 8.1668767929],
-        ["2020 / 11 / 03", "04:30", 2.8283495903, 4.103662014, 4.4018087387, 4.9847819805, 8.2466163635],
-        ["2020 / 11 / 03", "05:00", 2.9307193756, 4.1622138023, 4.5193009377, 5.113325119, 8.3263559341],
-        ["2020 / 11 / 03", "05:30", 3.0198979377, 4.2312819957, 4.4376957417, 5.17502141, 8.4372282028],
-        ["2020 / 11 / 03", "06:00", 3.1090764999, 4.3003501892000005, 4.3560905457, 5.236717701, 8.5481004715],
-        ["2020 / 11 / 03", "06:30", 2.9130270481, 4.3895642757, 4.5798389912, 5.3267338276, 8.878644466399997],
-        ["2020 / 11 / 03", "07:00", 2.7169775963, 4.4787783623, 4.8035874367, 5.416749954199999, 9.2091884613]
+        ["2020 / 11 / 03", "01:30", 2.4808146954, 3.7635395527, 4.2475838661, 4.9490222930000005, 7.6304917336]
     ])
+
+    metadata = {
+        "description": "forecast",
+        "database_id": "JERA Forecast",
+        "record_id": "test",
+        "type": "raw_data",
+        "path": "/opt/pydtk/test/records/jera/test.csv",
+        "content-type": "text/csv",
+        "contents": {
+            "test": {
+                "tags": [
+                    "test1",
+                    "test2"]
+            }
+        }
+    }
 
     if os.path.isfile(csv_path):
         os.remove(path=csv_path)
 
-    metadata_path = 'test/records/jera/test.csv.json'
-    with open(metadata_path, 'r') as fd:
-        metadata = json.load(fd)
-
     writer = BaseFileWriter()
     writer.write(data=data, metadata=metadata)
+
+    reader = BaseFileReader()
+    timestamps, new_data, columns = reader.read(metadata=metadata)
+    np.testing.assert_array_equal(data, new_data)
 
 
 @pytest.mark.extra
